separate routes for perpetual and spot tradingMode
